% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACF.R
\name{ACF}
\alias{ACF}
\alias{ACF}
\alias{CCF}
\alias{PACF}
\title{Auto- and Cross- Covariance and -Correlation Function Estimation}
\usage{
ACF(x, lag.max = NULL, type = c("correlation", "covariance", "partial"),
  plot = TRUE, na.action = na.fail, demean = TRUE, drop_lag0 = TRUE,
  ...)

CCF(x, y, lag.max = NULL, type = c("correlation", "covariance"),
  plot = TRUE, na.action = na.fail, ...)

PACF(x, lag.max = NULL, plot = TRUE, na.action = na.fail, ...)
}
\arguments{
\item{x}{a univariate or multivariate (not \code{ccf}) numeric time
    series object or a numeric vector or matrix, or an \code{"acf"} object.}

\item{lag.max}{maximum lag at which to calculate the acf.
    Default is \eqn{10\log_{10}(N/m)}{10*log10(N/m)} where \eqn{N} is the
    number of observations and \eqn{m} the number of series.  Will
    be automatically limited to one less than the number of observations
    in the series.}

\item{type}{character string giving the type of acf to be computed.
    Allowed values are
    \code{"correlation"} (the default), \code{"covariance"} or
    \code{"partial"}.  Will be partially matched.}

\item{plot}{logical. If \code{TRUE} (the default) the acf is plotted.}

\item{na.action}{function to be called to handle missing
    values. \code{na.pass} can be used.}

\item{demean}{logical.  Should the covariances be about the sample
    means?}

\item{drop_lag0}{drop the first lag?}

\item{\dots}{further arguments to be passed to \code{\link[=xyplot.acf]{xyplot.acf()}}.}

\item{y}{a univariate or multivariate (not \code{ccf}) numeric time
    series object or a numeric vector or matrix, or an \code{"acf"} object.}
}
\value{
If \code{plot = TRUE}, returns and plots a trellis object. Otherwise,
an \code{acf} object is returned
}
\description{
The following are versions of the functions detailed in
\code{\link[stats:acf]{stats::acf()}}. Usage is identical except for the plot functionality,
which you can read about in \code{\link[=xyplot.acf]{xyplot.acf()}}.
}
\details{
The confidence interval plotted in plot.acf is based on an uncorrelated
series and should be treated with appropriate caution. Using \code{ci.type = "ma"}
may potentially be less misleading.
}
\section{Functions}{
\itemize{
\item \code{ACF}: Autocorrelation and autocovariance

\item \code{CCF}: Crosscorrelation and crosscovariance

\item \code{PACF}: Partial autocorrelation and autocovariance
}}

\examples{
ACF(lh)
PACF(lh)
CCF(mdeaths, fdeaths)
}
\seealso{
\code{\link[=xyplot.acf]{xyplot.acf()}}, \code{\link[stats:acf]{stats::acf()}}, \code{\link[lattice:xyplot]{lattice::xyplot()}}
}
\author{
Original: Paul Gilbert, Martyn Plummer. Extensive modifications and
univariate case of pacf by B. D. Ripley. Adaptation to lattice by
Johan Larsson.
}
