% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma_acf.R
\name{arma_acf}
\alias{arma_acf}
\title{Compute (and Plot) Theoretical ACF for an ARMA Process}
\usage{
arma_acf(ar = numeric(), ma = numeric(), lag.max = NULL, pacf = FALSE,
  plot = FALSE, drop_lag0 = TRUE, ...)
}
\arguments{
\item{ar}{numeric vector of AR coefficients}

\item{ma}{numeric vector of MA coefficients}

\item{lag.max}{integer.  Maximum lag required.  Defaults to
    \code{max(p, q+1)}, where \code{p, q} are the numbers of AR and MA
    terms respectively.}

\item{pacf}{logical.  Should the partial autocorrelations be returned?}

\item{plot}{Set to \code{TRUE} to plot the resulting ACF via \link[lattice:xyplot]{lattice::xyplot}}

\item{drop_lag0}{Set to \code{TRUE} to drop the lag at 0.}

\item{\dots}{Graphical parameters passed onto \code{\link[lattice:xyplot]{lattice::xyplot()}}.}
}
\value{
A vector of (partial) autocorrelations, named by the lags or, if
\code{plot = TRUE}, a \code{trellis} object.
}
\description{
A wrapper around \code{\link[stats:ARMAacf]{stats::ARMAacf()}} that optionally plots the result
using a call to \code{\link[lattice:xyplot]{lattice::xyplot()}}.
}
\details{
The methods used follow Brockwell & Davis (1991, section 3.3).  Their
  equations (3.3.8) are solved for the autocovariances at lags
  \eqn{0, \dots, \max(p, q+1)}{0, \dots, max(p, q+1)},
  and the remaining autocorrelations are given by a recursive filter.
}
\examples{
arma_acf(ar = c(1.0, -0.25), ma = 1.0, lag.max = 10)
arma_acf(ar = -0.8, lag.max = 15, plot = TRUE)
}
\references{
Brockwell, P. J. and Davis, R. A. (1991) \emph{Time Series: Theory and
    Methods}, Second Edition.  Springer.
}
\seealso{
\code{\link[lattice:xyplot]{lattice::xyplot()}}, \code{\link[stats:arima]{stats::arima()}}, \code{\link[stats:ARMAtoMA]{stats::ARMAtoMA()}},
\code{\link[stats:acf2AR]{stats::acf2AR()}}.
}
