% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bug_fixes.R
\name{c.trellis}
\alias{c.trellis}
\alias{xyplot.list}
\title{Merge trellis objects, using same or different scales (fix)}
\usage{
\method{c}{trellis}(..., x.same = NA, y.same = NA, layout = NULL,
  merge.legends = FALSE, recursive = FALSE)

\method{xyplot}{list}(x, data = NULL, ..., FUN = lattice::xyplot,
  y.same = TRUE, x.same = NA, layout = NULL, merge.legends = FALSE)
}
\arguments{
\item{...}{ two or more trellis objects. If these are
    named arguments, the names will be used in the corresponding
    panel strips.
  }

\item{x.same}{ if \code{TRUE}, set the x scale relation to
    \code{"same"} and recalculate panel limits using data
    from all panels. Otherwise, the x scales in each panel
    will be as they were in the original objects
    (so in general not the same), the default behaviour.
  }

\item{y.same}{ as above, for y scales.
    Note that \code{xyplot.list} defaults to same y scales. Set to
    \code{NA} to leave them alone. }

\item{layout}{ value for \code{layout} of the new plot; see
    \code{\link{xyplot}}. }

\item{merge.legends}{ to keep keys or legends from all plots, not just
    the first. If multiple legends share the same \code{"space"}, they are
    packed together horizontally or vertically. }

\item{recursive}{ for consistency with the generic method, ignored. }

\item{x}{ a \code{list} of objects to plot individually, and then
    be combined into one final plot. }

\item{data}{ a lattice plot function, to be called on each element of
    the list \code{x}, along with \code{data} and \code{...} }

\item{FUN}{ a lattice plot function, to be called on each element of
    the list \code{x}, along with \code{data} and \code{...} }
}
\value{
a new trellis object.
}
\description{
This method is provided to fix a bug in \code{\link[latticeExtra:c.trellis]{latticeExtra::c.trellis()}}
that currently throws warnings if used. The function combines the panels of
multiple trellis objects into one.
}
\details{
This mechanism attempts to merge the panels from multiple trellis
  objects into one. The same effect could generally be achieved by
  either a custom panel function (where the display depends on
  \code{packet.number()}), or using \code{\link{print.trellis}}
  to display multiple trellis objects. However, in some cases it is more
  convenient to use \code{c()}. Furthermore, it can be useful to
  maintain the display as a standard lattice display, rather than a
  composite using \code{\link{print.trellis}}, to simplify
  further interaction.

  Many properties of the display, such as titles, axis settings
  and aspect ratio will be taken from the first object only.

  Note that combining panels from different types of plots does not
  really fit the trellis model. Some features of the plot may not work
  as expected. In particular, some work may be needed to
  show or hide scales on selected panels. An example is given below.

  Any trellis object with more than one conditioning variable will be
  "flattened" to one dimension, eliminating the multi-variate
  conditioning structure.
}
\seealso{
\code{\link{marginal.plot}} was the original motivating application,
  \code{\link{print.trellis}},
  \code{\link{update.trellis}},
  \code{\link{trellis.object}}
}
