% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dont_plot}
\alias{dont_plot}
\title{Suppress Plotting}
\usage{
dont_plot(x, ...)
}
\arguments{
\item{x}{Object to call \code{\link[graphics:plot.default]{graphics::plot()}} on.}
}
\value{
Invisibly returns whatever \code{plot(x)} would normally returns, but
does not plot anything (which is the point).
}
\description{
Suppress Plotting
}
\keyword{internal}
