% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grid_wrap}
\alias{grid_wrap}
\title{Wrapper for grid.arrange}
\usage{
grid_wrap(x, layout = NULL)
}
\arguments{
\item{x}{List of trellis objects.}

\item{layout}{A layout matrix or vector specifying rows and columns}
}
\value{
A list of \code{trellis} objects.
}
\description{
Wrapper for grid.arrange
}
\keyword{internal}
