% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latticework_theme.R
\name{latticework_theme}
\alias{latticework_theme}
\title{latticework lattice theme}
\usage{
latticework_theme(font_size = 12, point_size = 8, name = "pdf",
  color = name != "postscript", ...)
}
\arguments{
\item{font_size}{Font size. Will affect other settings such as padding as well.}

\item{point_size}{Point size. Will affect other settings.}

\item{name}{ name of the device for which the setting is required, as
    returned by \code{.Device}
  }

\item{color}{ logical, whether the initial settings should be color or
    black and white.  Defaults to \code{FALSE} for postscript devices,
    \code{TRUE} otherwise.  Note that this only applies to the initial
    choice of colors, which can be overridden using \code{theme} or
    subsequent calls to \code{\link{trellis.par.set}} (and by arguments
    supplied directly in high level calls for some settings).
  }

\item{\dots}{Additional named options appended via \code{\link[utils:modifyList]{utils::modifyList()}}.}
}
\value{
A list of graphical parameters that for instance could be supplied
inside a call to \code{\link[lattice:xyplot]{lattice::xyplot()}} or set via
\code{\link[lattice:lattice.options]{lattice::lattice.options()}}.
}
\description{
A custom theme for lattice.
}
\details{
The theme currently modifies the default lattice theme so that
\itemize{
\item paddings (margins) are minimized,
\item axis tick lengths are halved,
\item fontface for axis labels is set to bold, and
\item title size is decreased \emph{slightly.}
}

This theme is continually updated and is likely to change. Colors, for
instance, are likely to be updated eventually.
}
\examples{
xyplot(speed ~ dist, data = cars, par.settings = latticework_theme())

}
