% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbleplot.R
\name{make_bubbles}
\alias{make_bubbles}
\title{Make Bubbles}
\usage{
make_bubbles(x, maxsize)
}
\arguments{
\item{x}{A numeric vector.}

\item{maxsize}{The max size (in cex) of the bubbles.}
}
\value{
A list with the new bubbles as well as pretty breakpoints along with
their respective cex values.
}
\description{
Map \code{z} to the area of bubbles.
}
\keyword{internal}
