% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternaryplot.R
\name{panel.2ddensity}
\alias{panel.2ddensity}
\title{Title}
\usage{
panel.2ddensity(x, y, n = 100, region = TRUE, lims = if (ternary) c(0, 1,
  0, sqrt(3)/2) else c(current.panel.limits()$xlim,
  current.panel.limits()$ylim), ternary = FALSE,
  alpha = trellis.par.get("regions")$alpha)
}
\arguments{
\item{density_alpha}{}
}
\description{
Title
}
