% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel-qqmathci.R
\name{panel.qqmathci}
\alias{panel.qqmathci}
\title{Panel Function for Q-Q Plot Confidence Intervals}
\usage{
panel.qqmathci(x, y = x, distribution = qnorm, probs = c(0.25, 0.75),
  qtype = 7, groups = NULL, ci = 0.95, ci_col = "grey90", ...,
  identifier = "qqmathci")
}
\arguments{
\item{x}{ The original sample, possibly reduced to a fewer number of
    quantiles, as determined by the \code{f.value} argument to
    \code{qqmath}
  }

\item{y}{
    an alias for \code{x} for backwards compatibility
  }

\item{distribution}{
    quantile function for reference theoretical distribution.
  }

\item{probs}{
    numeric vector of length two, representing probabilities.
    Corresponding quantile pairs define the line drawn.
  }

\item{qtype}{
    the \code{type} of quantile computation used in
    \code{\link{quantile}}
  }

\item{groups}{
    optional grouping variable.  If non-null, a line will be drawn for
    each group.
  }

\item{ci}{Confidence level}

\item{ci_col}{Fill color of the confidence bands}

\item{\dots}{Arguments passed to \code{\link[lattice:panel.superpose]{lattice::panel.superpose()}} and
\code{\link[lattice:panel.polygon]{lattice::panel.polygon()}}.}

\item{identifier}{
    A character string that is prepended to the names of grobs
    that are created by this panel function.
  }
}
\value{
Adds confidence levels to a trellis device such as that
created by \code{\link[lattice:qqmath]{lattice::qqmath()}}.
}
\description{
Panel function to go along with \code{\link[lattice:qqmath]{lattice::qqmath()}} and possibly
\code{\link[lattice:panel.qqmathline]{lattice::panel.qqmathline()}}. Adds filled confidence bands to the Q-Q-plot.
}
\details{
The function tries to figure out the density function counterpart to
that provided in the argument \code{distribution}.
}
\examples{
library(lattice)
qqmath(~ Sepal.Width | Species, data = iris, distribution = qunif,
       panel = function(...) {
         panel.qqmathci(...)
         panel.qqmathline(...)
         panel.qqmath(...)
       })
}
\seealso{
\code{\link[lattice:panel.qqmathline]{lattice::panel.qqmathline()}}, \code{\link[lattice:qqmath]{lattice::qqmath()}}, and
\code{\link[lattice:panel.qqmath]{lattice::panel.qqmath()}}.
}
\author{
Modified by Johan Larsson from \code{\link[lattice:panel.qqmathline]{lattice::panel.qqmathline()}} by
Deepayan Sarkar.
}
