% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternaryplot.R
\name{panel.ternaryplot.clip}
\alias{panel.ternaryplot.clip}
\title{Plot Region Clipping for Ternary Plots}
\usage{
panel.ternaryplot.clip(xl = current.panel.limits()$x,
  yl = current.panel.limits()$y, border = "transparent", col = if
  (background$col == "transparent") "#FFFFFF" else background$col)
}
\arguments{
\item{xl}{X axis limits}

\item{yl}{Y axis limits}

\item{border}{Border color}

\item{col}{Polygon fill}
}
\value{
Plots a layer inside a panel of a \code{lattice} plot.
}
\description{
Plot Region Clipping for Ternary Plots
}
