% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternaryplot.R
\name{panel.ternaryplot.scales}
\alias{panel.ternaryplot.scales}
\title{Axes and Labels for Ternary Plots}
\usage{
panel.ternaryplot.scales(xlab, ylab, zlab, xlab.default, ylab.default,
  zlab.default, at = seq.int(0, 1, by = 0.2), ...)
}
\arguments{
\item{xlab}{
    Labels, have to be lists. Typically the user will not manipulate
    these, but instead control this via arguments to \code{cloud}
    directly.
  }

\item{ylab}{
    Labels, have to be lists. Typically the user will not manipulate
    these, but instead control this via arguments to \code{cloud}
    directly.
  }

\item{zlab}{
    Labels, have to be lists. Typically the user will not manipulate
    these, but instead control this via arguments to \code{cloud}
    directly.
  }

\item{xlab.default}{ for internal use  }

\item{ylab.default}{ for internal use  }

\item{zlab.default}{ for internal use  }

\item{at}{
    deals with specification of colors when \code{drape = TRUE} in
    \code{\link{wireframe}}.  \code{at} can be a numeric vector,
    \code{col.regions} a vector of colors, and \code{alpha.regions} a
    numeric scalar controlling transparency.  The resulting behaviour is
    similar to \code{\link{levelplot}}, \code{at} giving the breakpoints
    along the z-axis where colors change, and the other two determining
    the colors of the facets that fall in between.
  }

\item{...}{Currently ignored.}

\item{Where}{to draw tick marks.}
}
\value{
Plots a layer inside a panel of a \code{lattice} plot.
}
\description{
Axes and Labels for Ternary Plots
}
