% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternaryplot.R
\name{panel.ternaryplot.xyplot}
\alias{panel.ternaryplot.xyplot}
\title{Ternary Plot Wrapper for lattice::xyplot}
\usage{
panel.ternaryplot.xyplot(x, y, z, subscripts, ...)
}
\arguments{
\item{x}{Numeric vector of values in the original space}

\item{y}{Numeric vector of values in the original space}

\item{z}{Numeric vector of values in the original space}

\item{subscripts}{see \code{\link[lattice:xyplot]{lattice::xyplot()}}.}

\item{...}{Arguments that are passed on to \code{\link[lattice:panel.xyplot]{lattice::panel.xyplot()}}.}
}
\value{
Plots a layer inside a panel of a \code{lattice} plot.
}
\description{
This mainly exists to enable users to string together their own ternary plot
functions.
}
