% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqmath_methods.R
\name{qqmath.zoo}
\alias{qqmath.zoo}
\title{Q-Q diagram for zoo time series}
\usage{
\method{qqmath}{zoo}(x, data = NULL, reference = TRUE, ci = TRUE, ...)
}
\arguments{
\item{x}{A \code{zoo} object}

\item{data}{Ignored}

\item{reference}{Plot a reference line via \link{panel.qqmathline}?}

\item{ci}{Plot confidence levels?}

\item{\dots}{Graphical parameters to pass on to \code{\link[=qqmath]{qqmath()}}.}
}
\value{
Plots and returns a \code{trellis} object.
}
\description{
Draw quantile-Quantile plots of a sample against a theoretical distribution,
possibly conditioned on other variables.
}
\examples{
library(zoo)
zoo_lh <- zoo(lh)
qqmath(zoo_lh)
}
\seealso{
\code{\link[=qqmath]{qqmath()}}, \code{\link[zoo:zoo]{zoo::zoo()}}, \code{\link[=panel.qqmathline]{panel.qqmathline()}}.
}
