% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{require_pkg}
\alias{require_pkg}
\title{Throw An Error if A Required Package Is Unavailable}
\usage{
require_pkg(pkg)
}
\arguments{
\item{pkg}{The required package}
}
\value{
An error if the package namespace is not available.
}
\description{
Throw An Error if A Required Package Is Unavailable
}
\keyword{internal}
