% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rescale}
\alias{rescale}
\title{Uniform Rescaling}
\usage{
rescale(x, new_min = 0, new_max = 1, old_min = min(x, na.rm = TRUE),
  old_max = max(x, na.rm = TRUE))
}
\arguments{
\item{x}{A numeric vector to rescale}

\item{new_min}{New min}

\item{new_max}{New max}

\item{old_min}{Old min}

\item{old_max}{Old max}
}
\value{
A rescaled version of \code{x}.
}
\description{
Uniform Rescaling
}
\keyword{internal}
