% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seq_pal}
\alias{seq_pal}
\title{Sequential palette helper.}
\usage{
seq_pal(n, bias = 1, space = "Lab", ...)
}
\arguments{
\item{n}{Number of colors to generate}

\item{...}{Stuff to pass on to \code{\link[grDevices:colorRamp]{grDevices::colorRampPalette()}}}
}
\description{
Divides the regions palette from lattice in half when it does not make
sense to have a diverging palette.
}
\keyword{internal}
