% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_key}
\alias{setup_key}
\title{Key Setup}
\usage{
setup_key(legend, key, default_key, fun, pos = c("right", "top", "bottom",
  "left"))
}
\arguments{
\item{legend}{a list of legends, usually the \code{$legend} slot of a trellis
object.}

\item{key}{A key specification, usually the user input.}

\item{default_key}{The default key specifications that may be overridden by
the user.}

\item{fun}{The function to draw the key, such as \code{\link[lattice:draw.colorkey]{lattice::draw.colorkey()}}.}

\item{pos}{Preferences for the position of the new key.}
}
\value{
The original \code{legend} object with the addition of the key defined by
\code{key}, \code{default_key}, and \code{fun}.
}
\description{
Try to setup a key while also dodging existing keys
}
\keyword{internal}
