% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternaryplot.R
\name{ternaryplot.formula}
\alias{ternaryplot.formula}
\title{Title}
\usage{
\method{ternaryplot}{formula}(x, data = NULL, xlab = NULL, ylab = NULL,
  zlab = NULL, allow.multiple = is.null(groups) || outer,
  outer = !is.null(groups), panel = panel.ternaryplot, groups = NULL,
  subset = TRUE,
  drop.unused.levels = lattice.getOption("drop.unused.levels"), ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
