% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelliswork.R
\docType{package}
\name{trelliswork-package}
\alias{trelliswork}
\alias{trelliswork-package}
\title{trelliswork: Extensions to Lattice}
\description{
This package aims to extend the functionality of lattice
and latticeExtra with new functions and methods that make it easier to
be wholly reliant on trellis graphics in R. It is not an official extension
to lattice nor related to any of its developers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jolars/trelliswork}
  \item Report bugs at \url{https://github.com/jolars/trelliswork/issues}
}

}
\author{
\strong{Maintainer}: Johan Larsson \email{johanlarsson@outlook.com}

}
\keyword{internal}
