% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updateList}
\alias{updateList}
\title{Update a List with User Input}
\usage{
updateList(x, val)
}
\arguments{
\item{x}{A list to be updated.}

\item{val}{Stuff to update \code{x} with.}
}
\value{
Returns an updated list.
}
\description{
Modified version of \code{\link[lattice:::updateList]{lattice::::updateList()}} which relies on
\code{\link[utils:modifyList]{utils::modifyList()}}.
}
\seealso{
\code{\link[utils:modifyList]{utils::modifyList()}}
}
\keyword{internal}
