% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyplot_Arima.R
\name{xyplot.Arima}
\alias{xyplot.Arima}
\title{Diagnostic Plots for Time-Series Fits with lattice}
\usage{
\method{xyplot}{Arima}(x, data = NULL, which = 1:4, max.lag = NULL,
  gof.lag = NULL, qq.aspect = "iso", na.action = na.pass, main = NULL,
  layout = NULL, ...)
}
\arguments{
\item{x}{a fitted time-series model}

\item{data}{ignored (only added for method consistency)}

\item{which}{which plots should be plotted?}

\item{max.lag}{number of lags to compute ACF for}

\item{gof.lag}{the maximum number of lags for a Portmanteau
    goodness-of-fit test}

\item{qq.aspect}{aspect of Q-Q plot (see \code{\link[lattice:qqmath]{lattice::qqmath()}})}

\item{na.action}{what to do about na.values when computing ACF?}

\item{main}{optional titles for the plots}

\item{layout}{either a numeric vector with (columns, rows) to use in the call
to \code{\link[gridExtra:grid.arrange]{gridExtra::grid.arrange()}}, or a layout matrix which will then be
passed as the \code{layout_matrix} in \code{grid.arrange()}.}

\item{\dots}{parameters to pass to \code{\link[=xyplot]{xyplot()}}.}
}
\value{
Plots a lattice plot and returns a \code{trellis} object.
}
\description{
Diagnostic plots modelled after \code{\link[stats:tsdiag]{stats::tsdiag()}} with some modifications
and corrections of p-values in the Box--Ljung test.
}
\examples{
fit <- arima(lh, order = c(1, 1, 0))
xyplot(fit, layout = c(2, 2))
xyplot(fit, which = c(1:2, 4), layout = rbind(c(1, 1), c(2, 3)))
}
\seealso{
\code{\link[stats:tsdiag]{stats::tsdiag()}}, \code{\link[stats:arima]{stats::arima()}}, \code{\link[lattice:xyplot]{lattice::xyplot()}},
\code{\link[gridExtra:grid.arrange]{gridExtra::grid.arrange()}}, \code{\link[stats:Box.test]{stats::Box.test()}}, \code{\link[=ACF]{ACF()}}.
}
