% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyplot-forecast.R
\name{xyplot.forecast}
\alias{xyplot.forecast}
\title{Title}
\usage{
\method{xyplot}{forecast}(x, data = NULL, include = NULL, pi = 95,
  pi_col = "grey75", ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
