% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xyz_to_xy}
\alias{xyz_to_xy}
\title{Convert a formula from z ~ x * y to y ~ x}
\usage{
xyz_to_xy(form)
}
\arguments{
\item{form}{Three-dimensional lattice formula.}
}
\value{
A two-dimensional lattice formula.
}
\description{
Convert a formula from z ~ x * y to y ~ x
}
\keyword{internal}
