% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_fun}
\alias{get_fun}
\title{Retrieve a Function by Name or Definition}
\usage{
get_fun(fun)
}
\arguments{
\item{fun}{Character or function.}
}
\description{
Retrieve a Function by Name or Definition
}
\keyword{internal}
