% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latticework.R
\docType{package}
\name{latticework-package}
\alias{latticework}
\alias{latticework-package}
\title{latticework: Extensions for lattice}
\description{
Extensions to the R package lattice.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jolars/latticework}
  \item Report bugs at \url{https://github.com/jolars/latticework/issues}
}

}
\author{
\strong{Maintainer}: Johan Larsson \email{johanlarsson@outlook.com}

}
\keyword{internal}
