% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbleplot.R
\name{panel.bubbleplot}
\alias{panel.bubbleplot}
\title{Panel Function for Bubble Plots}
\usage{
panel.bubbleplot(x, y, z, groups = NULL, subscripts, cex = NULL, ...)
}
\arguments{
\item{x, y}{ variables to be plotted in the scatterplot }

\item{z}{A numeric vector that areas of circles will be mapped to.}

\item{groups}{Grouping variable (see \code{\link[lattice:xyplot]{lattice::xyplot()}}).}

\item{subscripts}{A vector of indexes to specify which observation to plot.
Normally does not need to be provided by the user.}

\item{cex}{Is used internally and user settings will be ignored.}

\item{...}{Further arguments to pass to \code{\link[lattice:panel.xyplot]{lattice::panel.xyplot()}}.}
}
\value{
Plots a layer inside a panel of a \code{lattice} plot.
}
\description{
Panel Function for Bubble Plots
}
