% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternaryplot.R
\name{panel.ternaryplot}
\alias{panel.ternaryplot}
\title{Panel Function for Ternary Plots}
\usage{
panel.ternaryplot(x, y, grid = TRUE, labs, density = TRUE,
  region = density, density_alpha = 0.25, ...)
}
\arguments{
\item{x}{Numeric vector}

\item{y}{Numeric vector}

\item{grid}{Draw a reference grid.}

\item{labs}{A list of labels for the diagram, which may be in the form
of expressions.}

\item{\dots}{Arguments passed on to \code{\link[lattice:xyplot]{lattice::xyplot()}}.}

\item{z}{Numeric vector}

\item{subscripts}{See \code{\link[lattice:panel.xyplot]{lattice::panel.xyplot()}}.}
}
\value{
Plots a layer inside a panel of a \code{lattice} plot.
}
\description{
Panel Function for Ternary Plots
}
