% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternaryplot.R
\name{panel.ternaryplot.scales}
\alias{panel.ternaryplot.scales}
\title{Axes and Labels for Ternary Plots}
\usage{
panel.ternaryplot.scales(
  xlab,
  ylab,
  zlab,
  xlab.default,
  ylab.default,
  zlab.default,
  at = seq.int(0, 1, by = 0.2),
  ...
)
}
\arguments{
\item{xlab, ylab, zlab}{
    Labels, have to be lists. Typically the user will not manipulate
    these, but instead control this via arguments to \code{cloud}
    directly.
  }

\item{xlab.default}{ for internal use  }

\item{ylab.default}{ for internal use  }

\item{zlab.default}{ for internal use  }

\item{at}{Where to draw tick marks.}

\item{...}{Currently ignored.}
}
\value{
Plots a layer inside a panel of a \code{lattice} plot.
}
\description{
Axes and Labels for Ternary Plots
}
