% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqmath-zoo.R
\name{qqmath.zoo}
\alias{qqmath.zoo}
\title{Q-Q diagram for zoo time series}
\usage{
\method{qqmath}{zoo}(x, data = NULL, xlab = "Theoretical quantiles",
  ylab = "Sample quantiles", ref = TRUE, ci = TRUE, ...)
}
\arguments{
\item{x}{A \code{zoo} object}

\item{data}{Ignored}

\item{ref}{Plots a reference line via \link[lattice:panel.qqmathline]{lattice::panel.qqmathline}}

\item{ci}{Plots confidence levels via \link{panel.qqmathci}}

\item{\dots}{Parameters to pass on to \code{\link[lattice:qqmath]{lattice::qqmath()}}.}
}
\value{
Plots and returns a \code{trellis} object.
}
\description{
Draw quantile-Quantile plots of a sample against a theoretical distribution,
possibly conditioned on other variables.
}
\examples{
qqmath(zoo::zoo(lh))
}
\seealso{
\code{\link[lattice:qqmath]{lattice::qqmath()}}, \code{\link[zoo:zoo]{zoo::zoo()}}, \code{\link[lattice:panel.qqmathline]{lattice::panel.qqmathline()}}.
}
\author{
Original by Deepayan Sarkar. Method by Johan Larsson.
}
