% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tactile.R
\docType{package}
\name{tactile-package}
\alias{tactile}
\alias{tactile-package}
\title{tactile: New and Extended Plots, Methods, and Panel Functions for Lattice}
\description{
Tactile (an anagram of Lattice) extends the functionality of
the Lattice package by providing new high-level functions, methods for
existing functions, panel functions, and theme.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jolars/tactile}
  \item Report bugs at \url{https://github.com/jolars/tactile/issues}
}

}
\author{
\strong{Maintainer}: Johan Larsson \email{johanlarsson@outlook.com}

}
\keyword{internal}
