% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tactile.R
\docType{package}
\name{tactile-package}
\alias{tactile}
\alias{tactile-package}
\title{tactile: New and Extended Plots, Methods, and Panel Functions for 'lattice'}
\description{
This package, 'tactile' (an anagram of lattice), extends the
functionality of the 'lattice' package by providing new high-level functions,
methods for existing functions, panel functions, and theme.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jolars/tactile}
  \item Report bugs at \url{https://github.com/jolars/tactile/issues}
}

}
\author{
\strong{Maintainer}: Johan Larsson \email{johanlarsson@outlook.com}

Other contributors:
\itemize{
  \item Sarkar Deepayan (lattice) [contributor, copyright holder]
}

}
\keyword{internal}
