% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tactile.R
\docType{package}
\name{tactile-package}
\alias{tactile}
\alias{tactile-package}
\title{tactile: New and Extended Plots, Methods, and Panel Functions for 'lattice'}
\description{
Extensions to 'lattice', providing new high-level functions, methods for existing functions, panel functions, and a theme.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jolars/tactile}
  \item Report bugs at \url{https://github.com/jolars/tactile/issues}
}

}
\author{
\strong{Maintainer}: Johan Larsson \email{johanlarsson@outlook.com} (\href{https://orcid.org/0000-0002-4029-5945}{ORCID})

Other contributors:
\itemize{
  \item Deepayan Sarkar (lattice) [contributor, copyright holder]
  \item Brian Ripley (stats::plot.acf) [contributor]
}

}
\keyword{internal}
