% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternaryplot.R
\name{ternaryplot}
\alias{ternaryplot}
\alias{ternaryplot.matrix}
\title{Ternary Plot}
\usage{
ternaryplot(x, data = NULL, ...)

\method{ternaryplot}{matrix}(x, data = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
A ternary diagram is a triangular diagram that displays proportions of
three variables.
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: A matrix of exactly three columns, wherein the the
first column is treated as \code{z}, the second as \code{x}, and the third as \code{y}.
}}

\examples{
ternaryplot(Fertility ~ Agriculture * Catholic, data = swiss)
}
