% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelliswork.R
\docType{package}
\name{trelliswork-package}
\alias{trelliswork}
\alias{trelliswork-package}
\title{trelliswork: Extensions to Lattice}
\description{
This package extends lattice and latticeExtra by providing new
high-level functions, methods for existing high-level functions, panel
functions, and themes. The objective is to replace base plotting functions
in order to be wholly reliant on trellis graphics in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jolars/trelliswork}
  \item Report bugs at \url{https://github.com/jolars/trelliswork/issues}
}

}
\author{
\strong{Maintainer}: Johan Larsson \email{johanlarsson@outlook.com}

}
\keyword{internal}
