% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocorrelation.R
\name{xyplot.acf}
\alias{xyplot.acf}
\title{Plot Autocovariance and Autocorrelation Functions}
\usage{
\method{xyplot}{acf}(x, data = NULL, ci = 0.95, ci.col = "gray50",
  ci.type = c("white", "ma"), ...)
}
\arguments{
\item{x}{an object of class \code{"acf"}.}

\item{data}{only provided for method consistency and is ignored.}

\item{ci}{coverage probability for confidence interval.  Plotting of
    the confidence interval is suppressed if \code{ci} is zero or
    negative.}

\item{ci.col}{colour to plot the confidence interval lines.}

\item{ci.type}{should the confidence limits assume a white noise
    input or for lag \eqn{k} an MA(\eqn{k-1}) input?   Can be abbreviated.}

\item{\dots}{graphical parameters passed on to \code{\link[lattice:xyplot]{lattice::xyplot()}}.}
}
\value{
Returns and plots a \code{trellis} object.
}
\description{
This is a version of \code{\link[stats:plot.acf]{stats::plot.acf()}} designed for the special versions
of the autocorrelation and autocovariance functions from \code{latticework}.
}
\examples{
z <- ts(matrix(rnorm(400), 100, 4), start = c(1961, 1), frequency = 12)
autocf(z, ci.col = "black")
}
\seealso{
\code{\link[=autocf]{autocf()}}, \code{\link[=crosscf]{crosscf()}}, \code{\link[=pautocf]{pautocf()}}, \code{\link[lattice:xyplot]{lattice::xyplot()}},
\code{\link[stats:plot.acf]{stats::plot.acf()}}, \code{\link[stats:acf]{stats::acf()}}.
}
