% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyplot-forecast.R
\name{xyplot.forecast}
\alias{xyplot.forecast}
\title{Title}
\usage{
\method{xyplot}{forecast}(x, data = NULL, interval = TRUE,
  interval_col = trellis.par.get("reference.line")$col,
  interval_alpha = 0.5, ...)
}
\arguments{
\item{...}{}
}
\value{
An object of class \code{"trellis"}. The
  \code{\link[lattice:update.trellis]{update}} method can be used to
  update components of the object and the
  \code{\link[lattice:print.trellis]{print}} method (usually called by
  default) will plot it on an appropriate plotting device.
}
\description{
Title
}
