% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyplot-methods.R
\name{xyplot.lm}
\alias{xyplot.lm}
\title{Lattice plot diagnostics for lm classes}
\usage{
xyplot.lm(x, which = c(1:3, 5), main = list("Residuals vs Fitted",
  "Normal Q-Q", "Scale-Location", "Cook's distance", "Residuals vs Leverage",
  "Cook's dist vs Leverage"), id.n = 3,
  labels.id = names(stats::residuals(x)), cex.id = 0.75,
  cook.levels = c(0.5, 1), label.pos = c(4, 2), ...)
}
\arguments{
\item{x}{A lm or glm object}

\item{which}{Which plots to plot}

\item{main}{Titles for the plots}

\item{id.n}{The number of extreme values to highlight}

\item{labels.id}{Labels for the extreme values.}

\item{cex.id}{Size of labels for extreme values.}

\item{cook.levels}{Cook's distance cutoffs.}

\item{label.pos}{Positions for id labels.}

\item{\dots}{Arguments to be passed to xyplot.}
}
\value{
Return a gtable list of grobs
}
\description{
Lattice plot diagnostics for lm classes
}
\examples{
library(lattice)
library(latticework)
fit <- lm(Sepal.Length ~ Sepal.Width, data = iris)
xyplot(fit)

}
